O3-4676: Patient forms app should render htmlforms without requiring explicit htmlform configuration in O3
