+++ packages/esm-patient-forms-app/src/forms/forms-dashboard.test.tsx
@@ -11,7 +11,7 @@ const mockUseConfig = jest.mocked(useConfig<ConfigObject>);
 const mockUseVisitOrOfflineVisit = useVisitOrOfflineVisit as jest.Mock;
 
 jest.mock('../hooks/use-forms', () => ({
-  useForms: jest.fn().mockReturnValueOnce({
+  useForms: jest.fn().mockReturnValue({
     data: [],
     error: null,
     isValidating: false,
